src/config.py
import os
from dotenv import load_dotenv
from sqlalchemy import create_engine

load_dotenv()

DB_USER = os.getenv("DB_USER", "postgres")
DB_PASSWORD = os.getenv("DB_PASSWORD", "admin")
DB_HOST = os.getenv("DB_HOST", "localhost")
DB_PORT = os.getenv("DB_PORT", "5432")
DB_NAME = os.getenv("DB_NAME", "stock_db")

DATABASE_URL = f"postgresql://{DB_USER}:{DB_PASSWORD}@{DB_HOST}:{DB_PORT}/{DB_NAME}"


engine = create_engine(DATABASE_URL, isolation_level="AUTOCOMMIT")

if __name__ == "__main__":
    try:
        with engine.connect() as conn:
            print("✅ Successfully connected toPostgreSQL!")
    except Exception as e:
        print(f"❌ Error connecting to PostgreSQL: {e}") 