requirements.txt
pandas>=2.2.0
yfinance>=0.2.36
psycopg2-binary>=2.9.9
SQLAlchemy<2.0.0
python-dotenv>=1.0.1
requests>=2.31.0
openai>=1.10.0
jupyter>=1.0.0
google-genai>=0.1.0